+++ packages/web/src/modules/content-stream/routes/route.utils.ts
@@ -1,5 +1,4 @@
 import { profileRoute } from '@/modules/profiles/routes/profile-route.util';
-import { IContent } from '@lyvely/common';
 
 export function contentRoute(pid: string, cid: string) {
   return profileRoute(`/stream/${cid}`, pid);