+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0.alpha3] - WIP
+
+### Fixed
+- Modal autofocus was not working due to new dialog
+- Center modal title on mobile
+
+### Changed
+- Initially hide main container scrollbar on small devices
+- Removed margins from filter navigation
+
+### Added
+- Implemented mobile footer menu
 
 ## [0.1.0.alpha2] - 2023-05-11
 ### Fixed